reviewApp
ReviewApp is a simple React Native application for managing and displaying user reviews. The app features a clean and intuitive interface, allowing users to add, view, and navigate through reviews. It demonstrates the use of React Navigation, Formik for form handling, Yup for validation, and custom components for a modular and maintainable codebase.